chore: correct enum imports
