+++ src/enum/EnumSuccessMessages.ts
@@ -1,0 +1,28 @@
+export enum EnumSuccessMessages {
+    // User success messages
+    USER_CREATED = 'Usuário criado com sucesso!',
+    USER_UPDATED = 'Usuário atualizado com sucesso!',
+    USER_DELETED = 'Usuário deletado com sucesso!',
+  
+    // Tutor success messages
+    TUTOR_CREATED = 'Tutor criado com sucesso!',
+    TUTOR_UPDATED = 'Tutor atualizado com sucesso!',
+    TUTOR_DELETED = 'Tutor deletado com sucesso!',
+    PHOTO_UPDATED = 'Foto atualizada com sucesso!',
+    PERSONAL_DATA_UPDATED = 'Dados pessoais atualizados com sucesso!',
+  
+    // Student success messages
+    STUDENT_CREATED = 'Aluno criado com sucesso!',
+    STUDENT_UPDATED = 'Aluno atualizado com sucesso!',
+    STUDENT_DELETED = 'Aluno deletado com sucesso!',
+  
+    // Lesson success messages
+    LESSON_CREATED = 'Aula criada com sucesso!',
+    LESSON_UPDATED = 'Aula atualizada com sucesso!',
+    LESSON_DELETED = 'Aula deletada com sucesso!',
+    LESSON_STATUS_UPDATED = 'Status da aula atualizado com sucesso!',
+  
+    // Authentication success messages
+    LOGIN_SUCCESS = 'Login realizado com sucesso!',
+    LOGOUT_SUCCESS = 'Logout realizado com sucesso!',
+  }
